Move elastc serializer to elastic2
